+++ pkg/iam/support_test.go
@@ -557,3 +557,193 @@ func TestIAM_ListStates(t *testing.T) {
 		})
 	}
 }
+
+func TestIAM_ListAccountSwitchKeys(t *testing.T) {
+	tests := map[string]struct {
+		params           ListAccountSwitchKeysRequest
+		responseStatus   int
+		expectedPath     string
+		responseBody     string
+		expectedResponse *ListAccountSwitchKeysResponse
+		withError        func(*testing.T, error)
+	}{
+		"200 OK with specified client": {
+			params: ListAccountSwitchKeysRequest{
+				ClientID: "test1234",
+			},
+			responseStatus: http.StatusOK,
+			expectedPath:   "/identity-management/v3/api-clients/test1234/account-switch-keys",
+			responseBody: `
+[
+  {
+    "accountName": "Test Name A",
+    "accountSwitchKey": "ABC-123"
+  },
+  {
+    "accountName": "Test Name A",
+    "accountSwitchKey": "ABCD-1234"
+  },
+  {
+    "accountName": "Test Name B",
+    "accountSwitchKey": "ABCDE-12345"
+  }
+]
+`,
+			expectedResponse: &ListAccountSwitchKeysResponse{
+				AccountSwitchKey{
+					AccountName:      "Test Name A",
+					AccountSwitchKey: "ABC-123",
+				},
+				AccountSwitchKey{
+					AccountName:      "Test Name A",
+					AccountSwitchKey: "ABCD-1234",
+				},
+				AccountSwitchKey{
+					AccountName:      "Test Name B",
+					AccountSwitchKey: "ABCDE-12345",
+				},
+			},
+		},
+		"200 OK without specified client": {
+			params:         ListAccountSwitchKeysRequest{},
+			responseStatus: http.StatusOK,
+			expectedPath:   "/identity-management/v3/api-clients/self/account-switch-keys",
+			responseBody: `
+[
+  {
+    "accountName": "Test Name A",
+    "accountSwitchKey": "ABC-123"
+  },
+  {
+    "accountName": "Test Name A",
+    "accountSwitchKey": "ABCD-1234"
+  },
+  {
+    "accountName": "Test Name B",
+    "accountSwitchKey": "ABCDE-12345"
+  }
+]
+`,
+			expectedResponse: &ListAccountSwitchKeysResponse{
+				AccountSwitchKey{
+					AccountName:      "Test Name A",
+					AccountSwitchKey: "ABC-123",
+				},
+				AccountSwitchKey{
+					AccountName:      "Test Name A",
+					AccountSwitchKey: "ABCD-1234",
+				},
+				AccountSwitchKey{
+					AccountName:      "Test Name B",
+					AccountSwitchKey: "ABCDE-12345",
+				},
+			},
+		},
+		"200 OK - no account switch keys": {
+			params: ListAccountSwitchKeysRequest{
+				ClientID: "test1234",
+			},
+			responseStatus:   http.StatusOK,
+			expectedPath:     "/identity-management/v3/api-clients/test1234/account-switch-keys",
+			responseBody:     `[]`,
+			expectedResponse: &ListAccountSwitchKeysResponse{},
+		},
+		"200 OK with query param": {
+			params: ListAccountSwitchKeysRequest{
+				ClientID: "test1234",
+				Search:   "Name A",
+			},
+			responseStatus: http.StatusOK,
+			expectedPath:   "/identity-management/v3/api-clients/test1234/account-switch-keys?search=Name+A",
+			responseBody: `
+[
+  {
+    "accountName": "Test Name A",
+    "accountSwitchKey": "ABC-123"
+  },
+  {
+    "accountName": "Test Name A",
+    "accountSwitchKey": "ABCD-1234"
+  }
+]
+`,
+			expectedResponse: &ListAccountSwitchKeysResponse{
+				AccountSwitchKey{
+					AccountName:      "Test Name A",
+					AccountSwitchKey: "ABC-123",
+				},
+				AccountSwitchKey{
+					AccountName:      "Test Name A",
+					AccountSwitchKey: "ABCD-1234",
+				},
+			},
+		},
+		"404 not found": {
+			params: ListAccountSwitchKeysRequest{
+				ClientID: "test12344",
+			},
+			responseStatus: http.StatusNotFound,
+			expectedPath:   "/identity-management/v3/api-clients/test12344/account-switch-keys",
+			responseBody: `
+{
+	"instances": "",
+    "type": "/identity-management/error-types/2",
+    "status": 404,
+    "title": "invalid open identity",
+	"detail": ""
+}				
+`,
+			withError: func(t *testing.T, err error) {
+				want := &Error{
+					Type:       "/identity-management/error-types/2",
+					Title:      "invalid open identity",
+					StatusCode: http.StatusNotFound,
+				}
+				assert.True(t, errors.Is(err, want), "want: %s; got: %s", want, err)
+			},
+		},
+		"500 internal server error": {
+			params: ListAccountSwitchKeysRequest{
+				ClientID: "test12344",
+			},
+			responseStatus: http.StatusInternalServerError,
+			responseBody: `
+{
+	"type": "internal_error",
+    "title": "Internal Server Error",
+    "detail": "Error making request",
+    "status": 500
+}`,
+			expectedPath: "/identity-management/v3/api-clients/test12344/account-switch-keys",
+			withError: func(t *testing.T, err error) {
+				want := &Error{
+					Type:       "internal_error",
+					Title:      "Internal Server Error",
+					Detail:     "Error making request",
+					StatusCode: http.StatusInternalServerError,
+				}
+				assert.True(t, errors.Is(err, want), "want: %s; got: %s", want, err)
+			},
+		},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			mockServer := httptest.NewTLSServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+				assert.Equal(t, test.expectedPath, r.URL.String())
+				assert.Equal(t, http.MethodGet, r.Method)
+				w.WriteHeader(test.responseStatus)
+				_, err := w.Write([]byte(test.responseBody))
+				assert.NoError(t, err)
+			}))
+			client := mockAPIClient(t, mockServer)
+			users, err := client.ListAccountSwitchKeys(context.Background(), test.params)
+			if test.withError != nil {
+				test.withError(t, err)
+				return
+			}
+			assert.NoError(t, err)
+			assert.Equal(t, test.expectedResponse, users)
+		})
+	}
+}